Large refactoring effort - switch to a new error-reporting system
Rather than exiting at the first error reached upon
elaboration/typechecking, we instead record the `Diagnostic`, and
continue checking as much as possible. This allows us to report multiple
errors across a single file

Should probably look into doing this for parsing as well.

A nice side benefit seems to be an increase in elaboration speed