Refactor sml-core/elaboration module to use arena allocation
* Most of the elaboration code had to be refactored to account for this,
but it's a relatively straightforward change, since this mostly involved
adding lifetimes and removing occurrences of `clone`

* Instead of splitting up the module into separate files, combine into a
single file (this made it a lot easier to incrementally refactor)

* Complete overhaul of internals of `TypeVar`. Rather than a shared
write-once mutable interior `Rc<UnsafeCell<Option<Type>>>`, we allocate
both `Type` and `TypeVar` in arenas, and we can simply share references!
That allows us to just use `Cell<Option<&'arena Type<'arena>>>`. This
also has the side benefit of simplifying "type sharing" from both an
implementation and conceptional standpoint

* Use specialized arena wrappers for Type so that we don't keep
allocating base types like `Unit` and `Exn`, etc. This could probably be
refactored in the future into a cleaner abstraction, but for now it
seems to work

* Add a dependency on the `typed-arena` crate for now, because there is
a bug (unsurprisingly) in the overly simplistic arena I implemented in 5
minutes, that led to some nasty debugging issues for type sharing.